requirements.txt
streamlit # The framework used to build the web application interface.
torch # PyTorch library for deep learning and tensor operations.
torchvision # PyTorch extension for computer vision tasks (datasets, transforms).
transformers # Hugging Face library providing the Vision Transformer (ViT) model.
pillow # Python Imaging Library for opening and manipulating image files.
